backend/user-service/user-service/src/analytics/serializers.py
from rest_framework import serializers

from .models import Event


class EventSerializer(serializers.ModelSerializer):
    class Meta:
        model = Event
        fields = ['id', 'event_type', 'user_id', 'metadata', 'created_at']
