backend/user-service/user-service/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
django-cors-headers>=3.10,<4.0
python-decouple>=3.5,<4.0
djangorestframework-simplejwt>=5.0,<6.0
africastalking>=1.0,<2.0
requests>=2.26,<3.0
Pillow>=9.0,<11.0
